Rename pkg/tun to pkg/vif
VIF is the commonly used abbreviation for Virtual Network
Interface and a more accurate description of this package.
