+++ pkg/vif/ip/v6header.go
@@ -6,7 +6,7 @@ import (
 
 	"golang.org/x/net/ipv6"
 
-	"github.com/telepresenceio/telepresence/v2/pkg/tun/buffer"
+	"github.com/telepresenceio/telepresence/v2/pkg/vif/buffer"
 )
 
 type V6Header []byte
